ManageUsers/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
importlib-resources==5.4.0
psycopg2-binary==2.9.3
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
